Keycloak Sync: Fix: keycloak_util import not working. Extra data added for SocialAccounts.
